feat: move config/database to app/core, fix imports, add initial Alembic migration
