+++ app/core/config.py
@@ -15,6 +15,7 @@ class Settings(BaseSettings):
     
     #Base de datos
     DATABASE_URL: str
+    DATABASE_URL_MIGRATION: str 
     POSTGRES_USER: str
     POSTGRES_PASSWORD: str
     POSTGRES_DB: str
