+++ app/core/security.py
@@ -7,7 +7,7 @@ from jose import JWTError, jwt
 from passlib.context import CryptContext
 from redis.asyncio import Redis
 
-from app.config import settings
+from app.core.config import settings
 from app.core.logging import get_logger
 
 logger = get_logger(__name__)
